requirements-dev.txt
ansible-core>=2.14.0
ansible-lint>=6.21.0
antsibull-docs
docker>=6.1.0
yamllint==1.32.0
pre-commit
-r requirements-test.txt
